40ignition-ostree: put growpart stamp file in /run
This stamp file was used to make sure coreos-growpart only ran on the
first boot when it ran in the real root. Nowadays, it runs in the
initramfs as part of `ignition-complete.target` before we even run
`ostree-prepare-root` (which meant we were actually creating the stamp
file in the initrd filesystem).

Having a stamp file is useful though for writing tests. So let's
repurpose the idea and put it in `/run` instead.