manifest: Include systemd-resolved for F33 change
This PR adds the systemd-resolved packages into FCOS. It is going to be used by
default in Fedora 33 per this change request:
https://fedoraproject.org/wiki/Changes/systemd-resolved

Adding the systemd-resolved packages back in gets us ready for this change.
It won't impact how DNS is configured or used and it lets users get ready for
the change in Fedora 33.