tests/kola: restrict DHCP propagation test to qemu
On cloud platforms, we have platform-specific code which takes
precedence of NTP-via-DHCP so this test won't work (specifically because
we set `PEERNTP=no`).

If there are other platforms which natively use DHCP for NTP setting, we
should enhance this test to handle those (so that e.g. we don't have to
set up dnsmasq at all).