20live: explicitly mount squashfs with `loop` option
This should be implied by the fact that we're trying to mount a
filepath, but let's be explicit. For comparison, this is how it's
documented in both squashfs-tools and in tldp:

https://github.com/plougher/squashfs-tools/blob/c570c6188811088b12ffdd9665487a2960c997a0/USAGE#L182

https://tldp.org/HOWTO/html_single/SquashFS-HOWTO/#creating

Spurred by RHBZ#1863466 where seemingly passing `-o loop` makes a
difference, though the root cause is likely somewhere else.