overlay: Install /usr/share/licenses/fedora-coreos-config/LICENSE
Thought about this after seeing
https://discussion.fedoraproject.org/t/fedora-core-os-redistribution/23134

Today fedora-coreos-config is special in that it's one of the only things
not an RPM.  Today one can use e.g. `rpm -qf /path/to/file` to find the
source RPM then map that to `/usr/share/licenses/$pkg/` to find the license,
but that's not true of stuff we ship here.

This doesn't solve the `rpm -qf` part, but it does put a file in the
expected place.