lockfiles: put in some overrides
Bodhi says some of these are stable already but the nightly composes
are failing so they aren't actually in the yum repos yet so let's
put them in as overrides for now.

Also update the generated lockfile with the output of a local run
of `cosa fetch --update-lockfile` to pre-seed file with F33 content.