ignition-ostree-rootfs: use our own tmpfs with size=80%
In f33 systemd, /run is now mounted with a smaller size cap instead of
the default 50%:

https://github.com/systemd/systemd/pull/15424

This was causing the rootfs reprovisioning code to run out of space.

Let's use this as an opportunity to be more explicit about the memory
semantics we want. Notably, we don't care about starving the host here;
either we have enough memory to reprovision the rootfs, or we should
fail. So let's just mount our own tmpfs with `size=80%`.

This in turn means that machines need even *less* RAM for rootfs
reprovisioning to work. A local test shows 2.5G seems to be enough.
Though let's keep recommending 4G in the docs to be safe.

A minor thing this also fixes is that we now clean up a leftover
directory and file from `/run` before we switchroot.

Closes: https://github.com/coreos/fedora-coreos-tracker/issues/644