manifests: fedora-coreos-base: add zram-generator
This was part of a F33 proposed change. We'll include the generator
but not the defaults subpackage because we don't want it enabled by
default just yet. We'll add docs for users instructing them how to
enable it by dropping down a file via Ignition/FCCT. The following
snippet is an example:

```
variant: fcos
version: 1.1.0
passwd:
  users:
    - name: core
      ssh_authorized_keys:
        - ssh-rsa AAAAB user@example.com
storage:
  files:
    - path: /etc/systemd/zram-generator.conf
      mode: 0644
      contents:
        inline: |
          # This config file enables a /dev/zram0 device with the default settings
          [zram0]
```

Closes https://github.com/coreos/fedora-coreos-tracker/issues/509