manifests: fedora-coreos-base: add back in podman-plugins, dnsmasq
We've found that there is some demand for host applications (podman,
NetworkManager, etc) using dnsmasq the binary for some various utility
operations. Let's add it back to the base so those host applications
can use the binary but let's mask the systemd service because we really
prefer if users are going to be hosting a DNS server for external
clients they use a container for that.

Closes https://github.com/coreos/fedora-coreos-tracker/issues/519