tests/kola: move swap-on-zram to misc-ign-ro
Let's add a new 'misc-ro' type test which is for testing things that is
solely configured via Ignition and doesn't need any "rw" changes at
runtime to test it. `swap-on-zram` is a good example of this, so let's
seed the test with that.

(cherry picked from commit 1f38b844d3b245d3a837edd98a2480475eb8d0a2)