Move buildroot image here
When using cosa to build OSes, we've been pretty good at shielding
ourselves from versioning differences between source and target
environments (for example, RHCOS is built by Fedora packages). But when
it comes to building things meant to run *in* that OS, we really should
match the buildroot of the target.

Right now, many of our projects use the buildroot cosa container to
build their projects in CI. But this doesn't necessarily match the
buildroot of the rest of the OS.

Since the buildroot to use is defined here (specifically via the
`releasever` key), let's move the concept of a "buildroot image" from
cosa to here. Specifically, we want a container image with all the deps
to the components we mostly hack on for use by both CI and local
developers.

This can then be used by CoreOS CI for building upstream projects, as
well as by cosa itself for building kolet.

See also: https://github.com/coreos/coreos-assembler/issues/1863