40ignition-ostree: copy BIOS-BOOT partition and boot sector
If the Ignition config creates any BIOS Boot partitions, save the
existing BIOS-BOOT contents and the corresponding boot sector before
ignition-disks (in case the partition is overwritten) and copy them to
the new partitions (and corresponding disks) afterward.  Also verify that
the offset of the new BIOS-BOOT partitions matches the old one, since
otherwise GRUB will fail when it tries to use them.  We don't require the
config to create a BIOS-BOOT RAID array because the OS doesn't use or
modify the BIOS-BOOT partition at runtime.