40ignition-ostree: copy PReP partition
If the Ignition config creates any PowerPC PReP partitions, save the
existing PReP contents before ignition-disks (in case the partition is
overwritten) and copy them to the new partitions afterward.  We don't
require the config to create a PReP RAID array because the OS doesn't
use or modify the PReP partition at runtime.