40ignition-ostree: be verbose about what we're mounting
We're sometimes hitting a flake in the RHCOS CI for LUKS-related tests:

    ignition-ostree-dracut-rootfs[909]: mount: /sysroot: unknown filesystem type 'crypto_LUKS'.
    systemd[1]: Failed to start Ignition OSTree: restore rootfs.

I think what may be happening here is that the `/dev/disk/by-label/root`
symlink is still pointing at the original partition which now hosts the
LUKS device and didn't actually get updated to point to the newly
created filesystem on top of it.

The `disks` stage does do a `udevadm trigger` on exit, though I'm not
sure how foolproof that is.

What we could do eventually to work around this is get the `device` from
the Ignition config and just mount that instead. Though to confirm first
this is what's happening, let's make the mount operations explicitly
call out what devices they're mounting.