Switch to sqlite rpmdb backend
There won't be any support for writing to the bdb backend in f34, so
e.g. pkglayering won't work (and obviously even composes wouldn't work
once we move cosa to f34).

All the production streams now have an f33 barrier release (as part of
the resolved workarounds), so it should be safe to switch to sqlite now.

For the other rpm-ostree-based variants, we'll probably just flip the
default in rpm-ostree to sqlite in f34. But at least in FCOS, we can do
this now so that we flush out any issues earlier and squash the ugly
warnings when querying the rpmdb.

Closes: https://github.com/coreos/fedora-coreos-tracker/issues/623