overlay: add warning if rootfs < 8G
We've recently had our first case of a "trapped" rootfs running out of
space for upgrades:

https://github.com/coreos/fedora-coreos-tracker/issues/731

Until we actually implement stronger behaviour for this, let's
explicitly check for this case and emit a warning if we detect it. In
the future, we'll look at making this a hard error by default (with an
escape hatch).

For more information, see:

https://github.com/coreos/fedora-coreos-tracker/issues/586#issuecomment-777220000