storage: tweak note re. rootfs sizing
Let's move it up so that it's right at the beginning of the document.
Mention that we will now warn and eventually fail provisioning if the
size is too small.

Also switch the FCCT to use the new `resize` key instead of the more
awkward `start_mib` on the following partition.