manifest: ship specific sssd subpackages instead
The `sssd` subpackage is a meta-package which just pulls in the default
set of packages. This default set changed in Fedora 34, which started
pulling in `python3-sssdconfig`:

https://bugzilla.redhat.com/show_bug.cgi?id=1927907

From the maintainer feedback there, it seems like it'd be better anyway
to be more explicit about the specific backends we want to ship. So
let's do that here: break out `sssd` into the client and the backends
for AD, IPA, KRB5, and LDAP.