manifest: use DEFAULT_HOSTNAME=localhost
systemd in f34 learned to read the default hostname from a
`DEFAULT_HOSTNAME` key in `/etc/os-release`.

Use that to set the hostname back to `localhost` on f34 so that once we
move over we can unpin the systemd build which had this same effect at
compilation time.

Ideally, this would be part of `fedora-release`, but I'm trying to aim
for a better fix where the compiled-in default hostname in Fedora is
back to `localhost` and only on the variants where we really want
`fedora` would we set `DEFAULT_HOSTNAME=fedora` in `fedora-release`.

So for now, let's just do this here.

Closes: #649