transposefs: error out cleanly if less than 1G of RAM available
I commonly do `cosa run` to test something related to rootfs
reprovisioning and forget to allocate more RAM. The failure mode is a
massive spam of "Input/output error" messages from `cp` unable to copy
more files into RAM (and there doesn't seem to be a way to have it fail
on first error; yet another reason we should rewrite this into !shell).

Instead, just sanity-check that we have at least 1G of RAM and error out
with a nice and clear message if not.