manifests: improve ability to share downstream
This introduces two new categories of packages which exist in both
the FCOS and RHCOS manifests.

The `system-configuration` manifest includes packages related to
configuration and management of the system itself.  Think storage,
NTP, users/groups, etc.

The `user-experience` manifest includes packages that will generally
make a users time on the system more pleasant.  Think commonly used
Linux utilities, container tooling, remote access, etc.

The goal of this change is to improve the ability to share these
manifests with RHCOS and reduce the duplication of packages in the
FCOS and RHCOS manifests.