overlay: rename coreos-growpart to ignition-ostree-growfs
Now that coreos-growpart is only in the initramfs, its scope is
essentially reduced to handling the `/sysroot`. Let's formalize this by
renaming the script to match the service which calls it and dropping
support for passing the target path as an argument.

Also freshen up the comments at the top of the file.