overlay/growfs: fix in-place root reprovisioning detection
The semantic we want to capture for auto-growing the rootfs is that we
want it to kick in as long as either no rootfs reprovisioning happened,
or it was in-place (i.e. reused the same partition).

We detected this by checking that the block device (i.e. partition)
backing the rootfs and its size didn't change. But this breaks with
in-place LUKS reprovisioning because the immediate block device does
change (to the LUKS container), even though it's ultimately still backed
by the same untouched partition.

Fix this by (1) verifying that the rootfs is still ultimately backed by
the same partition, and (2) that partition has not been resized.

This doesn't actually add support for growing LUKS containers yet, but
it's prep for that.

(cherry picked from commit 43437b494db35b764c89ff65130f32f505a4f7c9)