overlay/growfs: support growing LUKS containers
We need to support in-place LUKS root reprovisioning. This is pretty
straightforward, modulo the terrible clevis hack used to resize the
container. We'll want to circle back and clean this up once there's a
better interface to do this (but at least until then, we have tests
which cover this).

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1934174