overlay: temporarily ship `seq` in initrd
Previously we pinned clevis to v15 in order to work around
https://github.com/latchset/clevis/pull/295.

But overrides are not set up in the tagger for mechanical streams, so we
can't easily pin in `branched` and `rawhide`. Let's just unpin and
temporarily ship `seq` ourselves to work around this. Put that temporary
hack in `40ignition-ostree` since that's the closest module conceptually
to clevis.

(cherry picked from commit ad5eb32f8bde4d09bfc4320166ed5a23953d4d33)