buildroot: Drop USER creation, add createrepo_c+ostree
The USER bit breaks the ability to use this in Github actions.
Further, it's not useful for OpenShift because the platform
allocates a dynamic UID today (Eventually Kube user namespaces
will augment this).

We added this for the Jenkins/CCI, but previously we
were running the buildroot as uid 0 (i.e. Docker default) anyways.
If we want to enhance the flow to run as non-root, I think the
right thing is to do it in the platform and not the container.