overlay.d/35coreos-network: don't bring up NM if not needed
In dracut 053+ dracut starting running NetworkManager from a systemd
unit if systemd was used to start the system [1]. That systemd unit
will always get started if there is any NM configuration. Let's make
it only get started if we actually need the network. This is a
simplified version of the more complicated proposed fix upstream [2].

[1] https://github.com/dracutdevs/dracut/commit/c17c5b7
[2] https://github.com/dracutdevs/dracut/pull/1347

Fixes: https://github.com/coreos/fedora-coreos-tracker/issues/782