35coreos-live/coreos-livepxe-rootfs: retry fetching rootfs forever
We don't know how long it will take for networking to come up, so let's
match Ignition's semantic of retrying to fetch remote resources forever.

`curl` doesn't natively have support for this, so wrap it in a `while`
loop. This also works around the fact that el8's `curl` doesn't support
`--retry-all-errors`.

This was previously added in #929.