live-generator: don't call `man bootup`
Even though it's on a comment line, because it's in a heredoc, bash does
try to execute this. This fails on FCOS thankfully because there is no
`man` on FCOS, but it still logs an error message. (And... any derived
system which does ship `man`, I think this actually would dump the
manpage into the unit.)

(cherry picked from commit 3f19340ab9c3c9ec35cf79a89a6f2925f12e217c)