manifest.yaml: stop using -next repos
Both testing and next are on the same release now, so we can go back to
the stable repos for next-devel.

That stream isn't being built right now, but remove-graduated-overrides
still wants to try to check it, and I'd rather let it than disable it
for next-devel to be consistent and reduce churn whenever we flip the
switch. It's currently failing because those repos are no longer valid
for f34.

Also, it's possible we re-enable the stream anyway for rolling out some
of the systemd-oomd and swap suggestions.