overlay/05core: Set systemd LogColor=false
We deploy on servers.  The ANSI colors codes in the console
output often end up in e.g. log files where they both
make the logs hard to read *and* they make log files
invalid UTF-8.

This file needs to be created in both the real root and
the initramfs, so extend our `99journald-conf` to be
`99systemd-conf`.