ci/buildroot: bump nocache hack to defeat Quay.io caching
The images are being built, but the contents at this point are almost a
month out of date. If there's no sustainable way to work around this, we
should probably just move it to e.g. openshift/release and have it push
to Quay.io.

See: https://github.com/coreos/fedora-coreos-tracker/issues/890