35coreos-multipath: add Before=initrd-parse-etc.service in coreos-propagate-multipath-conf.service
Otherwise, we'll end up racing with `initrd-cleanup.service` which wants
to kill everything. It has `After=initrd.target` and we do have
`Before=initrd.target`, but that's not being respected, we think because
`initrd-parse-etc.service` does an explicit `systemctl start` on it.

Anyway, we need to dig more into this, but for now this will unblock us.