Switch to ostree-format: "oci"
(Take 2, now that we have https://github.com/coreos/coreos-assembler/pull/2417 )

Part of https://github.com/coreos/fedora-coreos-tracker/issues/812

In this initial step, we're merely switching the internal
tarball to be a different format.

A future step will change the FCOS pipeline to automatically
push this container to quay.io.