35coreos-ignition: allow setting kargs in live system if it's a no-op
In a live system, we can't set kargs in an Ignition config because we
don't control the bootloader settings.  If the config specifies kargs
anyway, we currently fail with "/dev/disk/by-label/boot: not a block
device", which isn't very clear.

In the live case, notice if the shouldExist kargs exist and the
shouldNotExist kargs do not, because then we don't need to do anything and
can succeed anyway.  Otherwise, fail with a clearer error message.

Fixes https://github.com/coreos/fedora-coreos-tracker/issues/917.