manifests: workaround F37 GPG key issue for rawhide
We'll workaround the issue by removing the F37 key from the list of
gpgkeys to load. We'll also add a test that fails if the f37 entry
isn't put back in place and we'll snooze that test. This gives us
a reminder to followup on this issue if we don't get it resolved
in upstream libdnf/rpm in the immediate future.

https://github.com/coreos/fedora-coreos-tracker/issues/925