Add Kubernetes e2e node tests
This patch adds new external tests which run the Kubernetes e2e node
tests. This will ensure that we don't inadvertedly break Kubernetes,
which is now itself tested on Fedora CoreOS.

The approach taken here is to model it as an external test. This makes
it easy for anyone to run it locally, which is important for
maintenance.

Note that these tests do *not* run by default: for now, I don't think we
should run these tests on every PR. They take a lot of resources and
time and it's not clear yet how flaky they'll be. Instead, we'll start
off by running them in the pipeline as a follow-on job. But as a next
step, I'd like to enable opt-in PR testing as well as testing as part of
the lockfile bumper.

One thing worth noting that makes it very different from a regular
external test is that the test binaries are built in the VM itself. This
is because (1) we may need newer golang than we have available in cosa,
and (2) cloning and building Kubernetes multiple times (per tested
release) is heavy and takes enough space that it's nicer to have it be
lifecycled to the VM itself, where it can be resource-constrained and
e.g. parallelized easily. For faster iteration, one can always just do
the same thing the test script does to build on the host and point the
k8s test harness at a pre-provisioned host.

Requires: https://github.com/coreos/coreos-assembler/pull/2505
Requires: https://github.com/kubernetes/kubernetes/pull/105764