manifests: add afterburn-dracut
The Fedora afterburn source package currently outputs an afterburn
and afterburn-dracut subpackage. In the process of unifying spec files
across Fedora/RHEL/C9S, we'd like to do the same for RHCOS. Before
splitting, we need to make sure that RHCOS won't lose the dracut bits.
So lower afterburn-dracut to a shared manifest.

An alternative is to have the afterburn package pull in afterburn-dracut
which will avoid any changes for RHCOS, but in theory the afterburn subpackage
could be useful without the dracut bits.