tests: Rework 'find' calls with multiple file type filters
Fedora has a recent enough `find` command to be able to use the `-type
f,d` option but that is not yet supported by `find` on RHEL.

We can not use `-type f -type d` instead as this filters on entries
being both a file and a directory which never happens. Thus we must
explicitely duplicate the entire filters and actions and use `-o` to OR
them.

This reworks the tests using that option to be compatible on both FCOS
and RHCOS. We can drop this change once RHCOS moves to RHEL 9.

This also includes `/usr` in the file-directory-permissions test and
tries to figure out which package the offending files/dires come from.

Fixes: https://github.com/coreos/fedora-coreos-config/commit/dac1f6835e1e730409a4062c6562ece32bddaee1