tests: use net.ifnames=0 for karg network tests
In order to allow the tests to work on both x86_64/BIOS, x86_64/UEFI
and aarch64/UEFI we use net.ifnames=0 so we can use eth1 in all
instances and it just works. Otherwise on x86_64/BIOS we get ens5
and on UEFI we get enp0s4.

Fixes https://github.com/coreos/fedora-coreos-tracker/issues/1060